+++ Motion_controller/src/test_tp_comm/src/test_frame/test_tool_get_all_valid_summary_info.cpp
@@ -79,7 +79,7 @@ int main()
     cout << "Reply : msg.error_code.data = " << recv_msg.error_code.data << endl;
 
     cout << "Reply : msg.data.tool_summary_info_count = " << recv_msg.data.tool_summary_info_count << endl;
-    for (int i = 0; i != recv_msg.data.tool_summary_info_count; ++i)
+    for (size_t i = 0; i != recv_msg.data.tool_summary_info_count; ++i)
     {
         cout << "Reply : msg.data.tool_summary_info[" << i << "].id = " << recv_msg.data.tool_summary_info[i].id << endl;
         cout << "Reply : msg.data.tool_summary_info[" << i << "].name = " << recv_msg.data.tool_summary_info[i].name << endl;
