+++ Motion_controller/src/test_tp_comm/src/test_io_manager/get_di.cpp
@@ -51,7 +51,7 @@ int main(int argc, char* argv[])
     msg.data.data[1] = atoi(argv[2]);
     msg.data.data[2] = atoi(argv[3]);
 
-    for (int i = 0; i != msg.data.data_count; ++i)
+    for (size_t i = 0; i != msg.data.data_count; ++i)
     {
         printf("msg.data.data[%d] = %d\n", i, msg.data.data[i]);
     }
