+++ Motion_controller/src/test_tp_comm/src/test_motion_control/axis_group/joints_feed_back_get.cpp
@@ -80,7 +80,7 @@ int main()
     cout << "Reply : msg.success = " << recv_msg.data.data << endl;
     cout << "Reply : msg.data.data_count = " << recv_msg.data.data_count << endl;
 
-    for (int i = 0; i != recv_msg.data.data_count; ++i)
+    for (size_t i = 0; i != recv_msg.data.data_count; ++i)
         cout << "Reply : msg.data.data["<< i <<"] = " << recv_msg.data.data[i] << endl;
 
     usleep(200000);
