+++ Motion_controller/src/test_tp_comm/src/test_motion_control/axis_group/soft_limit_get_manu.cpp
@@ -81,11 +81,11 @@ int main()
     cout << "Reply : msg.error_code.data = " << recv_msg.error_code.data << endl;
     cout << "Reply : msg.limit.positive_limit.data_count = " << recv_msg.limit.positive_limit.data_count << endl;
     cout << "Reply : msg.limit.negative_limit.data_count = " << recv_msg.limit.negative_limit.data_count << endl;
-    for (int i = 0; i != recv_msg.limit.positive_limit.data_count; ++i)
+    for (size_t i = 0; i != recv_msg.limit.positive_limit.data_count; ++i)
     {
         cout << "Reply : msg.limit.positive_limit.data[" << i << "] = " << recv_msg.limit.positive_limit.data[i] << endl;
     }
-    for (int i = 0; i != recv_msg.limit.negative_limit.data_count; ++i)
+    for (size_t i = 0; i != recv_msg.limit.negative_limit.data_count; ++i)
     {
         cout << "Reply : msg.limit.negative_limit.data[" << i << "] = " << recv_msg.limit.negative_limit.data[i] << endl;
     }
